scripts/requirements.txt
pymupdf>=1.24.0
qdrant-client>=1.13.2
openai>=2.0.0
python-dotenv>=1.0.1
